# docs/swagger.yaml
definitions:
...
  shortpress-server_internal_api.IDListResponseData-string:
    properties:
      hasMore:
        description: Whether there are more items
        example: true
        type: boolean
      items:
        description: ID information list
        items:
          type: string
        type: array
      page:
        description: Current page number
        example: 1
        type: integer
      pageSize:
        description: Items per page
        example: 20
        type: integer
      total:
        description: Total number of videos
        example: 100
        type: integer
    type: object
...
  shortpress-server_internal_api.Response:
    properties:
      code:
        type: integer
      data: {}
      info:
        type: string
    type: object
...
  shortpress-server_internal_api.SiteAddPlaylistsRequest:
    properties:
      playlistIds:
        description: Playlist ID array
        example:
        - '["playlist-1"'
        - '"playlist-2"]'
        items:
          type: string
        type: array
      siteId:
        description: Site ID, CHAR(36)
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
    required:
    - playlistIds
    - siteId
    type: object
  shortpress-server_internal_api.SiteCreateResponse:
    properties:
      code:
        type: integer
      data:
        type: string
      info:
        type: string
    type: object
  shortpress-server_internal_api.SiteDelPlaylistsRequest:
    properties:
      playlistIds:
        example:
        - '["playlist-1"'
        - '"playlist-2"]'
        items:
          type: string
        type: array
      siteId:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
    required:
    - playlistIds
    - siteId
    type: object
  shortpress-server_internal_api.SiteGetResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/shortpress-server_internal_api.SiteInfo'
      info:
        type: string
    type: object
  shortpress-server_internal_api.SiteInfo:
    properties:
      domain:
        example: example.com
        type: string
      facebookPixelId:
        description: Facebook Pixel ID, optional
        example: "1234567890"
        type: string
      googleAnalyticsId:
        description: Google Analytics ID, optional
        example: UA-123456789-1
        type: string
      logo:
        example: http://example.com/logo_updated.png
        type: string
      name:
        example: MySite Updated
        type: string
      officialDomain:
        $ref: '#/definitions/shortpress-server_internal_types.OfficialDomain'
      path:
        description: Site path, VARCHAR(255), required
        example: /home
        type: string
      redirect:
        example: false
        type: boolean
      seo:
        $ref: '#/definitions/shortpress-server_internal_api.SiteSeo'
      seoMultiLang:
        description: 多语言配置，JSON格式，不做处理，直接存储
        items:
          type: integer
        type: array
      siteId:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      siteMultiLang:
        description: 多语言配置，JSON格式，不做处理，直接存储
        items:
          type: integer
        type: array
      status:
        example: 1
        type: integer
      templateId:
        example: tpl_01f3b2
        type: string
      templateName:
        description: Template name
        example: Modern Template
        type: string
      thinkingdataAppId:
        description: Thinking Data App ID, optional
        example: app-1234567890
        type: string
    type: object
...
  shortpress-server_internal_api.UserChangeStatusRequest:
    properties:
      email:
        description: |-
          Email is the user's email address
          Required: true
          Format: email
        type: string
      siteId:
        description: |-
          SiteID is the unique identifier of the site
          Required: true
        type: string
      status:
        description: |-
          Status is the new status to assign to the user
          Required: true
          Enum: 2 3 127
          2: activate - User is active
          3: forbidden - User is forbidden
          127: delete - User is deleted
        type: integer
    required:
    - email
    - siteId
    - status
    type: object
...
  shortpress-server_internal_api.UserInfo:
    properties:
      createdAt:
        example: 1609459200
        type: integer
      email:
        example: user@example.com
        type: string
      lastLoginAt:
        example: 1609459200
        type: integer
      nickname:
        example: John Doe
        type: string
      premiumExpiresAt:
        description: Unix timestamp
        example: 1609459200
        type: integer
      premiumType:
        description: '0: free, 1: premium'
        example: 0
        type: integer
      status:
        example: 1
        type: integer
      updatedAt:
        example: 1609459200
        type: integer
      userId:
        example: user-123456
        type: string
    type: object
...
  shortpress-server_internal_api.VideoListResponseData:
    properties:
      items:
        description: Video list
        items:
          $ref: '#/definitions/shortpress-server_internal_api.VideoInfo'
        type: array
      page:
        description: Current page number
        example: 1
        type: integer
      pageSize:
        description: Items per page
        example: 20
        type: integer
      total:
        description: Total count
        example: 100
        type: integer
    type: object
...
  shortpress-server_internal_types.OfficialDomain:
    properties:
      subdomain:
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: shortpress
    url: http://www.swagger.io/support
  description: This is shortpress server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Shortpress Server API
  version: "0.8"
paths:
...
  /api/site/add-playlists:
    post:
      consumes:
      - application/json
      description: Add one or more video playlists to the target site.
      parameters:
      - description: Bearer user token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Add playlists parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/shortpress-server_internal_api.SiteAddPlaylistsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully added playlists response
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.Response'
      security:
      - Bearer: []
      summary: Add playlists to site
      tags:
      - site
  /api/site/batch-get:
    get:
      consumes:
      - application/json
      description: Batch query brief site information by site ID
      parameters:
      - description: Bearer user token
        in: header
        name: Authorization
        type: string
      - description: List of site IDs, comma-separated
        example: '"site-123456,site-123457"'
        in: query
        name: siteIds
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Returns a list of brief site information
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.VideoListResponseData'
      security:
      - Bearer: []
      summary: Batch get site information
      tags:
      - site
  /api/site/create:
    post:
      consumes:
      - application/json
      description: Create a site and its default SEO information. Optionally pass
        templateId to initialize pages from a template and save into site_builder_data.
      parameters:
      - description: Bearer user token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Create site parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/shortpress-server_internal_api.SiteInfo'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully created site response
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.SiteCreateResponse'
      security:
      - Bearer: []
      summary: Create site and default SEO information
      tags:
      - site
  /api/site/get:
    get:
      consumes:
      - application/json
      description: Query basic site information and corresponding SEO information
        by site ID
      parameters:
      - description: Bearer user token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Unique site ID
        example: '"123e4567-e89b-12d3-a456-426614174000"'
        in: query
        name: siteId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Returns site and SEO information
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.SiteGetResponse'
      security:
      - Bearer: []
      summary: Query site and SEO information
      tags:
      - site
  /api/site/list:
    get:
      consumes:
      - application/json
      description: Paginate query for all site information owned by the current user,
        including basic site information and status
      parameters:
      - description: Bearer user token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Returns site list data
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.IDListResponseData-string'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.Response'
      security:
      - Bearer: []
      summary: Get user's site list
      tags:
      - site
  /api/site/modify:
    post:
      consumes:
      - application/json
      description: Modify site information and SEO configuration, siteId is a required
        field.
      parameters:
      - description: Bearer user token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Modify site parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/shortpress-server_internal_api.SiteInfo'
      produces:
      - application/json
      responses:
        "200":
          description: Returns whether modification was successful
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.Response'
      security:
      - Bearer: []
      summary: Modify site and SEO information
      tags:
      - site
...
  /api/site/path/valid:
    get:
      consumes:
      - application/json
      description: Check if a site path is valid and available for use
      parameters:
      - description: Bearer user token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Site path to validate
        example: '"my-site"'
        in: query
        name: path
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Path is valid and available
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.Response'
      security:
      - Bearer: []
      summary: Validate site path availability
      tags:
      - site
  /api/site/remove-playlists:
    post:
      consumes:
      - application/json
      description: Remove one or more video playlists from the target site.
      parameters:
      - description: Bearer user token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Remove playlists parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/shortpress-server_internal_api.SiteDelPlaylistsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully removed playlists response
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.Response'
      security:
      - Bearer: []
      summary: Remove playlists from site
      tags:
      - site
  /api/site/user/change/status:
    post:
      consumes:
      - application/json
      description: Change a user's status (activate, forbidden, delete) within a site.
        If deleted, the user is truly removed.
      parameters:
      - description: Bearer user token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Parameters for changing user status
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/shortpress-server_internal_api.UserChangeStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Status changed successfully
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.Response'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.Response'
      security:
      - Bearer: []
      summary: Change user status
      tags:
      - site
  /api/site/user/info:
    get:
      consumes:
      - application/json
      description: Retrieve user information for a specific email within a site
      parameters:
      - description: Bearer user token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Site ID
        example: '"site-123456"'
        in: query
        name: siteId
        required: true
        type: string
      - description: User email
        example: '"user@example.com"'
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Returns user information
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.UserInfo'
      security:
      - Bearer: []
      summary: Get user information by email
      tags:
      - site
  /api/site/user/list:
    get:
      consumes:
      - application/json
      description: Retrieve users associated with a site with pagination, filtering,
        and sorting
      parameters:
      - description: Bearer user token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Site ID
        example: '"site-123456"'
        in: query
        name: siteId
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        maximum: 100
        minimum: 1
        name: pageSize
        type: integer
      - description: Search term for email or nickname
        example: '"john"'
        in: query
        name: query
        type: string
      - default: 1
        description: 'Sort type: 0=login time desc, 1=created time desc, 2=login time
          asc, 3=created time asc'
        in: query
        name: sortType
        type: integer
      - default: -1
        description: 'Status filter: -1=all, 1=inactive, 2=active, 3=disabled, 4=deleted'
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Returns list of users
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.UserListResponseData'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.Response'
      security:
      - Bearer: []
      summary: Get users for a specific site
      tags:
      - site
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"